chore: Refactor CLI
