+++ censys/cli/commands/__init__.py
@@ -1,7 +1,10 @@
 """Censys CLI commands."""
-from .config import cli_asm_config, cli_config
-from .hnri import cli_hnri
-from .search import cli_search
-from .view import cli_view
+from . import asm, config, hnri, search, view
 
-__all__ = ["cli_asm_config", "cli_config", "cli_hnri", "cli_search", "cli_view"]
+__all__ = [
+    "asm",
+    "config",
+    "hnri",
+    "search",
+    "view",
+]
